Make non-u variants primary.
This doesn't affect what code type checks but does affect tools
producing docs from these stubs.
