+++ typeshed/stdlib/uerrno.pyi
@@ -1,26 +1,1 @@
-from typing import Mapping
-
-errorcode: Mapping[int, str]
-
-EACCES: int
-EADDRINUSE: int
-EAGAIN: int
-EALREADY: int
-EBADF: int
-ECONNABORTED: int
-ECONNREFUSED: int
-ECONNRESET: int
-EEXIST: int
-EHOSTUNREACH: int
-EINPROGRESS: int
-EINVAL: int
-EIO: int
-EISDIR: int
-ENOBUFS: int
-ENODEV: int
-ENOENT: int
-ENOMEM: int
-ENOTCONN: int
-EOPNOTSUPP: int
-EPERM: int
-ETIMEDOUT: int
+from errno import *
